extract components for clarity
